+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Version 0.0.12
 
-  * 
+  * Attribute Section.myscout\_email\_address\_send defaults to an empty String
+  * Attribute Section.myscout\_email\_address\_copy defaults to an empty String
+  * Attribute Section.myscout\_email\_address\_send renamed to myscout\_email\_address\_from
 
 ## Version 0.0.11